Fixing a regression with integration of two controllers
